Create OCR Simulator
This simulator runs a simulated OCR process that steps through the
Query, Observation, Report steps to produce reports destined for a
contract. This simulator does not commit anything and does not sign
observations or reports.

Run a simulation of multiple nodes querying the underlying chain data
set in the context of OCR epochs and rounds.