Update benchmarks to account for multiple nodes building reports
More nodes on an OCR network results in more observations sent to
all nodes. This increases the workload for each node and likewise
the latency for the Report operation. This commit steps the number
of nodes/observations for each benchmark test to display the latency
expected as the number of nodes increases.