Report performance improvement
The 16 node benchmark scenario was reduced from 300ms to just over
100ms to complete as well as flattening out the latency difference
between the total number of nodes. The culprit was in using json
encoding/decoding in plase of gob since the encoding/decoding
workloads are small for this case.

The change in encoding method also added some performance gains for
the Observation benchmarks, though minimal.