Fix context cancellation for parallel upkeep check
When sending jobs to the worker group, the context from the parent
function 'parallelCheck' was not being added to the work item itself.
This would cause the checkUpkeep to continue processing when the
result listener had already exited.

This change passes the context from the calling function to each work
item such that when that context is cancelled, the work item terminates
allowing the results listener to finish processing the results and
terminating the function cleanly with no lingering items in the worker
group results channel.