Refactor InFlight State
In processing incoming logs, delete items in in-flight state.

When checking upkeeps, return state as in-flight if the key exists.

The registry should return the block number returned from the contract
instead of the block number used to check the upkeep. This keeps the
cache keys and perform data wrapper in sync.