Update upkeep service to 'Lockout' concept
To make the service more readable and separate responsibilities from
OCR and the check upkeep service, in-flight was changed to lockout and
new functions were added to the upkeep service.

This solution is an intermediary between the current service and another
future refactor.

The performed report log was also changed from a subscriber model to
a polling model which should directly support the log poller.