Add mutex to ensure race doesn't happen
The CL node race tests were flagging the worker results aggregator as
a potential for race conditions. This shouldn't happen in production
since the wait group is coordinating the workers with the results
aggregator, but a mutex was added so that all value changes would be
synchronized to ensure a race condition won't happen.