Modify observation context before sampler
The observation context can cancel potentially causing all results from
sampling to be lost for that node in that round. Modifying the parent
context to reduce the deadline allows a window for all workers to
shut down and report as well as enough time for the observation to
complete in time with at least some results.

This allows an observation to not be a total loss for that round if
RPC calls are taking longer than expected.