Worker Group Refactor
An issue was discovered where a running worker group can result in a hanging
process when called to stop. This was due to outstanding work items still in
queue not being executed where the work items contained wait group items.

The solution was to drain the queue on stop and return a service context error
as a result for each job. This should allow the calling function to terminate
gracefully.

Additionally, workers no longer send results to a channel. Instead, results are
sent to the worker group via a callback function where the worker group adds the
result to an array of results and then notify the caller that work is done. This
eliminates the chance that a process could hang due to an incorrectly sized
results queue.