Add an UpkeepObservation type  (#72)
* Add a new UpkeepObservation type

- Stores a single block key and a list of upkeep identifiers
- Emitted by the observation phase and consumed by the report phase
- Dedupe calculates the median block from all upkeep observations
- Dedupe constructs new upkeep keys using the median block and zeroth upkeep ID

* Update tests

* Update OCR tests

* Key fields in tests

* Add a LatestBlock number to the service

- Regenerate mocks
- Update tests

* Added observation and reporting limitations

* Rewrite limitedLengthEncode

* Refactor shuffledDedupedKeyList into shuffleUniqueObservations

* Addressing comments

* Added reportBlockLag to the offchain config

* Added tests

* Close of Closed Channel in Worker Group

The `Stop()` method closes a channel, which causes all workers in the
worker group to exit. The finalizer also attempts to close the same
channel.

The finalizer always runs during a garbage collection cycle when no
more references exist to the WorkerGroup. If the `Stop()` function is
called first, the finalizer will attempt to close a closed channel.

The solution is to wrap the contents of the `Stop()` function in a
sync.Once and call the stop function from the finalizer instead of
closing the channel directly.

* Get block key from upkeep sampling, remove purging of results

* Fix failing tests

* Rename GetActiveUpkeepKeys to GetLatestActiveUpkeepKeys

* small fix in maing and logs

* fix keys variable name and lint

* rename functions for clarity

* Minor updates and cleanup

* Refactor GetLatestActiveUpkeepKeys to GetActiveUpkeepIDs

* Fix sample size check

* Log error on decode error

* Fix typo

* Change the BlockKey type to an interface (#76)

* Change the BlockKey type to an interface

* Minor fixes

* Fixes

* use after in filter

* add clarity in filter naming

* address comment to simplify block key implementation

* fix tests

* fix test

---------

Co-authored-by: Roman Behma <begmaroman@gmail.com>
Co-authored-by: Akshay Aggarwal <akshay.aggarwal@smartcontract.com>

* rename KeyBlockNumber to CheckBlockNumber for clarity

* change json key values

* update comment

* cleanup unused function

* Add logging for non cached keys

* add mroe debugging logging

* cleanup logs

* add TODOs

* Worker Group Refactor

An issue was discovered where a running worker group can result in a hanging
process when called to stop. This was due to outstanding work items still in
queue not being executed where the work items contained wait group items.

The solution was to drain the queue on stop and return a service context error
as a result for each job. This should allow the calling function to terminate
gracefully.

Additionally, workers no longer send results to a channel. Instead, results are
sent to the worker group via a callback function where the worker group adds the
result to an array of results and then notify the caller that work is done. This
eliminates the chance that a process could hang due to an incorrectly sized
results queue.

* ignore nancy dependency vul

* cleanup unused registry function

* validate all observations

* add test

* revert cleanup

---------

Co-authored-by: Roman Behma <begmaroman@gmail.com>
Co-authored-by: Akshay Aggarwal <akshay.aggarwal@smartcontract.com>
Co-authored-by: Awbrey Hughlett <awbrey.hughlett@smartcontract.com>