Chore/flatten project files to pkg (#109)
* Flatten OCR Keepers to /pkg

The goal is to move all types that could be used externally into
the /pkg package. This helps simplify the plugin surface area as
complex code isolation isn't really needed at this point.

* Separate Marshalling and Validation

Because the validator type was introduced and is an interface instead
of a concrete list of validations, the marshalling behavior and
validation needed to be separated. Likewise, the tests were separated.

To complete full validation testing, the individual validators should
be tested separately.

* upgrade tool version

* remove sim-v1 cmd in makefile

* fix linting errors