Fix memory leak
The worker group was creating new arrays and channels for each work set and
not deleting them when done. Memory was gradually being used up by empty
arrays and channels going unused.

The coordinator stop channel had a length of 0 and would block indefinitely
until the ability to be read. The length was changed to 1 to ensure the value
was sent over the channel.

The observer was automatically starting on creation with a finalizer that
stopped the service. Since we are moving to explicit starts and stops, these
should no longer be needed.