Fix flaky test
Using random bytes and casting to string was resulting in a flaky test because
the string was being left padded with zeros and cast once again to a string.
Random bytes combined with zeros can cause string casting to assume different
rune combinations that would result in a string length longer than the intended
number of bytes.