Log Trigger Pre-build Hook w/ Skeleton Plugin
To begin tying more pieces together, a new function was added to the log trigger
eligibility flow that allows it to function as a pre-build hook for the plugin.
This leads to more complex component stacking up to the plugin instance.

The result store was moved into the v3 package.

The plugin code was moved to a plugin package to avoid circular dependencies and
to isolate plugin specific code.