Build Hook Dependencies (#171)
Build hooks depend on various stores to work and currently they are being
passed to each build hook. This convolutes the concept of what the build
hooks are for in that that should be writers to the Observation.

Instead, the stores can be provided to each build hook which will simplify
the hook interface, make the build hook concept more extendable, and make
the intent of the hooks more explicit.