Complete Block Coordination Flow
Included in this commit is a complete setup for coordinating the latest common
block apart from adding the blocks to the outcome. The data flow begins with a
block history source and uses instructions in the OCR plugin to coordinate.

This makes an assumption that coordination will happen within the Outcome, but
that piece is still incomplete. After the assumed coordination, the latest
common block is added to the metadata store.

Still missing:
- block coordination in Outcome
- timer to set/reset instructions for coordination