Timed Recovery Flow
The retry/recovery tickers were eliminated in favor of a generic scheduled ticker
that groups items into a tick after some amount of wait time. The observer was
also made into a generic to allow different observer types that essentially
follow the same filter/process/post pattern with a ticker as a driver.

The primary goal of the refactoring was to create a different pathway for payloads
from retry such that recovery could receive different types of values from
different sources than retry.

The end result provides a generalized metadata store, a metadata postprocessor
for recovery, and a ticker/observer combination for recovery.