Outcome History Ring Buffer
Block coordination happens independently of other data coordination, but is a
required value for other processes. To correctly be able to merge coordinated
blocks and coordinated recoverable ids, outcome history is required. This can
either be local state or we can use the outcome to store previous coordinated
values.

This commit introduces a ring buffer of outcome history that is maintained
collectively by all nodes. This allows all nodes to coordinate on various bits
of data as well as recent history.