Correct Unmarshalling of Base Structs
The base automation ocr types contain interface values that lose their types
when marshalled. Since the plugin doesn't know what those types are at the time
of unmarshalling, they should be kept as raw json.

This will allow consumers of the base types to unmarshal the values independent
of the plugin.