Basic layout for conditional upkeep flow
This commit includes a structurally complete conditional eligibility flow and
integrates it into the plugin constructor. Also all known dependecies are
surfaced in the delegate config.

The internal functionality of the conditional upkeep flow is still incomplete.
The goal of this commit is to put the skeleton components in place to begin
connecting further components that complete this flow.