Wire up samples and recovery (#207)
* Wire up samples and recovery

This commit wires together the metadata store and OCR for conditional flows and
recovery flows. When samples are detected in an outcome, a coordinated block is
searched for in the outcome history. If one is found, the samples are merged
with the latest coordinated block and pushed to the runner.

The same process exists for recovery items. When recovery items are detected in
an outcome, a coordinated block is searched for in the outcome history. If one
is found, the recovery items are merged with the latest coordinated block and
pushed to the runner.

* added some test coverage

* addressed comments

* converted helper functions into member functions of the outcome and added lots more tests

* fix case of empty proposals

* fix case of empty samples