Add WorkID to structs
This commit adds a new property to UpkeepPayload, TransmitEvent, and
ReportedUpkeep. All three structs have an ID and should have a corresponding
WorkID to identify the upkeep at different stages of the pipeline.

Also included in this commit is removing CheckBlock from UpkeepPayload to help
avoid confusion over unused properties.