Refactor to Separate Components
Instead of a single simulated contract, the components were sorted into two
basic groups: loaders and trackers.

Loaders add transactions to blocks as they are produced. Loaders can selectively
add to a block if the block number reaches a threshold or for every block. Most
loaders are to inject configuration events at specific blocks, but also include
transmits.

Trackers follow specific event types in blocks and typically collect data. This
is similar to the separation of an RPC and blockchain in that they can be in
different states at the same time. One example of a tracker is for active
upkeeps; when upkeeps are created.