Ratio Returns Zero in Rounding (#267)
For selecting sample sizes, the goal is to split the work between multiple nodes
and mostly has been focused on larger numbers when sampling. In the case of a
sample ratio less than 0.5 and a count of 1, the value returned would be zero.

There aren't any cases where a zero value would be desireable and at the minimum
of 1 upkeep, all nodes could check the same upkeep without putting strain on any
node.

The fix for now is to return only values greater than zero. In the case of a
value being rounded down to zero, return 1 unless the incoming count is zero.