Proposal for Upkeep State Telemetry
Relying on log output for debugging has been challenging at best and error prone at
worst. This is frequently due to the fluctuating nature of full text log output as
changes occur in the code base. On top of that, log collection at a full debug level
for simple state telemetry is highly wasteful as most of the log output doesn't
directly relate to state changes.

This proposal introduces distinct states of a specific upkeep that are indicated as
being altered at specific points in the program. The approach replaces the default
logger with a special telemetry logger that adds telemetry logging functions.

To view potential output of the telemetry logging, run the following in a local
terminal:

$ make simulator && ./bin/simulator --simulate --verbose -f ./tools/simulator/plans/only_log_trigger.json

Then view `./simulation_plan_logs/simulation.log` for "Status Delays per Work ID". The
output attempts to indicated the time taken for an upkeep to reach specific states.
This allows for evaluation of the time spent at each point in the process.